Force mod update on each instance
With previous logic, if an instance shares a mod with another,
the first one would update the file, but the second would then
not realize a restart is necessary to use the latest version.

We must stop each server that requires a mod update, even if
another instance updates the mod for us.